HTMLSelectElement.item() does not behave according to specification
HTMLSelectElement.item() does not behave according to specification:
http://www.w3.org/TR/html51/forms.html#dom-select-item
http://dom.spec.whatwg.org/#dom-htmlcollection-item

Namely, we should not throw if the offset argument is negative. Also,
the argument should be mandatory.

IE10 and Firefox do not throw when HTMLSelectElement.item() is called
with a negative value (as per the specification).

The argument is mandatory in Firefox but optional in IE10 (it returns
null when called without argument). This looks like a bug in IE because
other item() getters' argument is usually mandatory in IE (e.g.
NodeList.item()).

BUG=237739

Review URL: https://chromiumcodereview.appspot.com/21834002

git-svn-id: svn://svn.chromium.org/blink/trunk@155515 bbb929c8-8fbe-4397-9dbb-9b2b20218538