Manually "restore" context, instead of using save/restore when RenderLayerBacking paints.
Assuming that RenderLayer painting code does not change the transform
state of the context, then RenderLayerBacking can avoid a potentially
costly context save/restore.  Instead, we can just manually undo the
one translation that was applied in when RenderLayerBacking paints,
to be equivalent to a "restore".

It is worth noting that we had already been relying on this assumption
for a while. Only recently in r155466 was this assumption removed.  But
this new code that uses save/restore may have caused a slight perf
regression.

BUG=268273

Review URL: https://chromiumcodereview.appspot.com/22177002

git-svn-id: svn://svn.chromium.org/blink/trunk@155543 bbb929c8-8fbe-4397-9dbb-9b2b20218538