Do not dispatch 'focus' or 'blur' event twice when window active state is changed.
* Bug
On Windows, 'focus' event for the window object or a focused element is
dispatched twice when the browser tab gets active, and 'blue' evnet is
dispatched twice when the browser tab gets inactive.
On OSX, 'blur' evnet is dispatched twice when the browser tab gets inactive.
On Android, we don't have this bug.

* Root cause
Focus/blue event is dispatched in dispatchEventsOnWindowAndFocusedNode
in these cases. FocusController::setFocused(bool) calls it, and
FocusController::setActive(bool) also calls it if setFocused(true) was
already called.

On Windows, these functions are called in the following sequence:
  Activation:
    WebView::setFocus(true)
      FocusController::setFocused(true)
      FocusController::setActive(true)
    WebView::setIsActive(true)
  Inactivation:
    WebView::setIsActive(false)
      FocusController::setActive(false)
    WebView::setFocus(false)
      FocusController::setFocused(false)

On OSX, these functions are called in the following sequence:
  Activation:
    WebView::setIsActive(true)
      FocusController::setActive(true)
    WebView::setFocus(true)
      FocusController::setFocused(true)
      FocusController::setActive(true)
  Inactivation:
    WebView::setIsActive(false)
      FocusController::setActive(false)
    WebView::setFocus(false)
      FocusController::setFocused(false);

On Android, these functions are called in the following sequence:
  Activation:
    WebView::setFocus(true)
      FocusController::setFocused(true)
      FocusController::setActive(true)
  Inactivation:
    WebView::setFocus(false)
      FocusController::setFocused(false);

* Fix
We don't need to call dispatchEventsOnWindowAndFocusedNode in
setActive() because we don't make a window focused state true when the
window is inactive.

BUG=117307

Review URL: https://chromiumcodereview.appspot.com/22076003

git-svn-id: svn://svn.chromium.org/blink/trunk@155558 bbb929c8-8fbe-4397-9dbb-9b2b20218538