Clean up title code.
Don't store a title on DocumentLoader, as it's largely redundant with Document.
Move a couple of state modifications to Document.

Add a dispatchDidReceiveTitle() call to the same document navigation case. We
have a special title update in RenderViewImpl in the same document navigtion
case, and it's the only reason we need to expose title information on
WebDataSource. By sending this title update call from core/, we can trim down
the WebDataSource API.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/22898002

git-svn-id: svn://svn.chromium.org/blink/trunk@155962 bbb929c8-8fbe-4397-9dbb-9b2b20218538