Renames to make the code easier to read
Rename FloatIntervalSearchAdapter to make it clear that it actually does
something other than search, since it has been non-obvious in the past
that calling it isn't a no-op.

Rename highValue and lowValue variables in the adapter to make it clearer
what they actually stand for.

Review URL: https://chromiumcodereview.appspot.com/22937003

git-svn-id: svn://svn.chromium.org/blink/trunk@156031 bbb929c8-8fbe-4397-9dbb-9b2b20218538