Reland "Clean up title code."
This reverts commit 001264754f744710523cef8c552efe57344463e3.

Don't store a title on DocumentLoader, as it's largely redundant with Document.
Move a couple of state modifications to Document.

Add a dispatchDidReceiveTitle() call to the same document navigation case. We
have a special title update in RenderViewImpl in the same document navigtion
case, and it's the only reason we need to expose title information on
WebDataSource. By sending this title update call from core/, we can trim down
the WebDataSource API.

BUG=

Review URL: https://chromiumcodereview.appspot.com/23086002

git-svn-id: svn://svn.chromium.org/blink/trunk@156051 bbb929c8-8fbe-4397-9dbb-9b2b20218538