Disable touch hit testing on overlay layers
The compositor hit testing doesn't know how to deal with layers that may be transparent to events, so for now we need to assume any touch on an overlay needs to be handled on the main thread.

This is exactly analogous to an earlier issue http://crbug.com/173547 where overlay layers needed to be marked as non-fast-scrolling.

BUG=178281

Review URL: https://chromiumcodereview.appspot.com/22674007

git-svn-id: svn://svn.chromium.org/blink/trunk@156067 bbb929c8-8fbe-4397-9dbb-9b2b20218538