Remove the Custom Element 'defined' state.
Refactoring. No behavior change (although less detailed asserts.)

Custom Elements processing used to depend on four states:

  1. Not a custom element
  2. An upgrade candidate (waiting for a definition)
  3. Defined (with a definition, but waiting for upgrade processing)
  4. Upgraded

This refactoring mushes the second and third states together. This is
in preparation for introducing a new state related to parsing.

Some of the structure (for example, consolidating more state setting
in CustomElementRegistrationContext; setting the WaitingForUpgrade
state relatively early in createCustomTagElement and setTypeExtension
instead of in didCreateUnresolvedElement) and naming
(WaitingForUpgrade instead of UpgradeCandidate) is in anticipation of
the introduction of the new state.

BUG=234509

Review URL: https://chromiumcodereview.appspot.com/23130005

git-svn-id: svn://svn.chromium.org/blink/trunk@156135 bbb929c8-8fbe-4397-9dbb-9b2b20218538