[CSS Grid Layout] Invalid cast in RenderGrid::addChild
RenderGrid::addChild wrongly assumed that new children were
necessarily boxes (and tripped on the ASSERT in toRenderBox).

Note that inline children will be wrapped into anonymous
wrappers which makes the ASSERT in removeChild correct, but
addChild is called on the DOM child, which can be arbitrary.
Also SVG is not an issue as <svg> is a RenderBox.

BUG=266906

Review URL: https://chromiumcodereview.appspot.com/23225003

git-svn-id: svn://svn.chromium.org/blink/trunk@156166 bbb929c8-8fbe-4397-9dbb-9b2b20218538