Implement computed style for animation shorthand and align parsing with the spec.
Implement the computed style for the animation shorthand which will avoid
to return "".

Improve parsing of the shorthand by adding the play-state property part of the parsing. This align with the spec and it was changed at the end of last
year.

Remove dead code in CSSParser-in.cpp inherited from the transitions.

There are few commented tests that are not passing as they do not yet
align with the spec. It's a trickier issue that I will fix in a following patch (bug is linked).

BUG=272963, 234612

Review URL: https://chromiumcodereview.appspot.com/23011009

git-svn-id: svn://svn.chromium.org/blink/trunk@156170 bbb929c8-8fbe-4397-9dbb-9b2b20218538