Unexpose -webkit-marquee-direction
This change introduces the concept of an internal CSS property that
is only allowed when parsing the UA and internal inline stylesheets.
The internal properties starts with -internal- and are not exposed
to the web (checked by our existing webexposed/ tests + a new one
to be sure).

To ensure that the previous logic works, -webkit-marquee-direction
was renamed to -internal-marquee-direction and thus unexposed from
the web platform.

BUG=274055

Review URL: https://chromiumcodereview.appspot.com/22849019

git-svn-id: svn://svn.chromium.org/blink/trunk@156256 bbb929c8-8fbe-4397-9dbb-9b2b20218538