Add a --no-android-logging flag to run-webkit-tests.
Currently we log every single adb command as part of --debug-rwt-logging
when running the tests on the android port. This makes the log very
verbose, and most of the time this is noise. This patch adds a way to
turn that logging off, though it is still on by default. Eventually we
should flip the default to have to turn the logging on (or get rid of
the logging altogether), but there's some clean up of logging that is
actually useful that should happen first.

R=peter@chromium.org, ojan@chromium.org, abarth@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/22980017

git-svn-id: svn://svn.chromium.org/blink/trunk@156406 bbb929c8-8fbe-4397-9dbb-9b2b20218538