Avoid some unneeded malloc in GridIterator
The issue was that the local variables didn't have the same type
(Vector<RenderBox*>) as that of the grid representation
(Vector<RenderBox*, 1>), which forced a hidden copy. The fix is
to add the missing typedefs to avoid type mismatchs in the future.

This caused PerformanceTests/Layout/auto-grid-lots-of-data.html to
severely regress after crrev.com/22867006. On a Z620, this improves
the test from 83.5 run/s to 88 run/s (a 5% improvement).

BUG=275745

Review URL: https://chromiumcodereview.appspot.com/23137008

git-svn-id: svn://svn.chromium.org/blink/trunk@156416 bbb929c8-8fbe-4397-9dbb-9b2b20218538