Ensure keySplines is valid in SMIL animations
This patch fixes a crash in SMIL animations when keySplines are not
specified. The SMIL spec is clear on this:
http://www.w3.org/TR/2001/REC-smil-animation-20010904/#AnimFuncCalcMode
"If there are any errors in the keyTimes specification (bad values,
too many or too few values), the animation will have no effect."

This patch simply checks that keyTimes is not empty. Previously,
splinesCount was set to be m_keySplines.size() + 1 in
SVGAnimationElement.cpp; this patch changes splinesCount to be equal
to m_keySplines.size() to make the logic easier to follow and to
match other checks in SVGAnimationElement::startedActiveInterval.

BUG=276111

Review URL: https://chromiumcodereview.appspot.com/23254006

git-svn-id: svn://svn.chromium.org/blink/trunk@156452 bbb929c8-8fbe-4397-9dbb-9b2b20218538