Implement support for unprefixed keyframes rules resolution.
This commit implements the support for parsing and resolving unprefixed
@keyframes. The behavior is as followed :
- For a given animation name if only a @-webkit-keyframe or @keyframe
is defined then we resolve it.
- For a given animation name if both @-webkit-keyframes and @keyframes
are defined, the unprefixed version is always taken.
- As before if you define multiple time a @keyframes for the same animation
name (prefixed or not) the last declaration is resolved (as before).

This commit doesn't implement yet the CSSOM part of the @keyframes, it
will be done in a separate patch.

Review URL: https://chromiumcodereview.appspot.com/23264017

git-svn-id: svn://svn.chromium.org/blink/trunk@156490 bbb929c8-8fbe-4397-9dbb-9b2b20218538