Fix a crash in impl side painting for image decoding
When image data is reciving very slowly there is a rare case that an
ImageDecoder is created twice and inserted into ImageDecodingStore.
This condition triggers a crash.

The following sequence is causing the problem:
1. ImageFrameGenerator tries to resume decoding, resulted an empty
   frame.
2. A little more data is received.
3. ImageFrameGenerator creates a new decoder and tries to insert it to
   cache.

This fix is to decode only once instead of resume decoding and then
decode again with a new decoder.

BUG=275421

Review URL: https://chromiumcodereview.appspot.com/22801010

git-svn-id: svn://svn.chromium.org/blink/trunk@156494 bbb929c8-8fbe-4397-9dbb-9b2b20218538